requirements.txt
discord.py
instaloader
pytube
yt-dlp
asyncpg
deep_translator
langdetect